Move truncate component
