+++ app/src/components/utility/Truncate.tsx
